Implement OAuth Facebook authentication.
