+++ src/main/java/com/intellias/intellistart/interviewplanning/config/SecurityCredentialsConfig.java
@@ -4,7 +4,6 @@ package com.intellias.intellistart.interviewplanning.config;
 import com.intellias.intellistart.interviewplanning.services.JwtTokenProvider;
 import com.intellias.intellistart.interviewplanning.services.UserService;
 import javax.servlet.http.HttpServletResponse;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Lazy;
@@ -24,77 +23,88 @@ import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 import org.springframework.web.filter.CorsFilter;
 
+/**
+ * Security credential config class.
+ */
 @EnableWebSecurity
 @EnableGlobalMethodSecurity(prePostEnabled = true)
 public class SecurityCredentialsConfig extends WebSecurityConfigurerAdapter {
 
-    private final UserDetailsService userDetailsService;
-    private final JwtConfig jwtConfig;
-    private final JwtTokenProvider tokenProvider;
-    private final UserService userService;
+  private final UserDetailsService userDetailsService;
+  private final JwtConfig jwtConfig;
+  private final JwtTokenProvider tokenProvider;
+  private final UserService userService;
 
-    @Autowired
-    @Lazy
-    public SecurityCredentialsConfig(UserDetailsService userDetailsService, JwtConfig jwtConfig,
-        JwtTokenProvider tokenProvider, UserService userService) {
-        this.userDetailsService = userDetailsService;
-        this.jwtConfig = jwtConfig;
-        this.tokenProvider = tokenProvider;
-        this.userService = userService;
-    }
+  /**
+   * SecurityCredentialsConfig class constructor.
+   */
+  @Autowired
+  @Lazy
+  public SecurityCredentialsConfig(UserDetailsService userDetailsService, JwtConfig jwtConfig,
+      JwtTokenProvider tokenProvider, UserService userService) {
+    this.userDetailsService = userDetailsService;
+    this.jwtConfig = jwtConfig;
+    this.tokenProvider = tokenProvider;
+    this.userService = userService;
+  }
 
 
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http
-            .cors().and()
-            .csrf().disable()
-            .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-            .and()
-            .exceptionHandling().authenticationEntryPoint(
-                (req, rsp, e) -> rsp.sendError(HttpServletResponse.SC_UNAUTHORIZED))
-            .and()
-            .addFilterBefore(
-                new JwtTokenAuthenticationFilter(jwtConfig, tokenProvider, userService),
-                UsernamePasswordAuthenticationFilter.class)
-            .authorizeRequests()
-            .antMatchers(HttpMethod.POST, "/facebook/signin").permitAll()
-            .anyRequest().authenticated();
-    }
+  @Override
+  protected void configure(HttpSecurity http) throws Exception {
+    http
+        .cors().and()
+        .csrf().disable()
+        .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+        .and()
+        .exceptionHandling().authenticationEntryPoint(
+            (req, rsp, e) -> rsp.sendError(HttpServletResponse.SC_UNAUTHORIZED))
+        .and()
+        .addFilterBefore(
+            new JwtTokenAuthenticationFilter(jwtConfig, tokenProvider, userService),
+            UsernamePasswordAuthenticationFilter.class)
+        .authorizeRequests()
+        .antMatchers(HttpMethod.POST, "/facebook/signin").permitAll()
+        .anyRequest().authenticated();
+  }
 
-    @Override
-    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-        // Configure DB authentication provider for user accounts
-        auth.userDetailsService(userDetailsService).passwordEncoder(passwordEncoder());
-    }
+  @Override
+  protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+    // Configure DB authentication provider for user accounts
+    auth.userDetailsService(userDetailsService).passwordEncoder(passwordEncoder());
+  }
 
-    @Bean
-    public BCryptPasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
+  @Bean
+  public BCryptPasswordEncoder passwordEncoder() {
+    return new BCryptPasswordEncoder();
+  }
 
-    @Bean(BeanIds.AUTHENTICATION_MANAGER)
-    @Override
-    public AuthenticationManager authenticationManagerBean() throws Exception {
-        return super.authenticationManagerBean();
-    }
+  @Bean(BeanIds.AUTHENTICATION_MANAGER)
+  @Override
+  public AuthenticationManager authenticationManagerBean() throws Exception {
+    return super.authenticationManagerBean();
+  }
 
-    @Bean
-    public CorsFilter corsFilter() {
-        final UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        final CorsConfiguration config = new CorsConfiguration();
-        config.setAllowCredentials(true);
-        config.addAllowedOrigin("*");
-        config.addAllowedHeader("*");
-        config.addAllowedMethod("OPTIONS");
-        config.addAllowedMethod("HEAD");
-        config.addAllowedMethod("GET");
-        config.addAllowedMethod("PUT");
-        config.addAllowedMethod("POST");
-        config.addAllowedMethod("DELETE");
-        config.addAllowedMethod("PATCH");
-        source.registerCorsConfiguration("/**", config);
-        return new CorsFilter(source);
-    }
+  /**
+   * corsFilter configuration.
+   *
+   * @return CorsFilter.
+   */
+  @Bean
+  public CorsFilter corsFilter() {
+    final UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+    final CorsConfiguration config = new CorsConfiguration();
+    config.setAllowCredentials(true);
+    config.addAllowedOrigin("*");
+    config.addAllowedHeader("*");
+    config.addAllowedMethod("OPTIONS");
+    config.addAllowedMethod("HEAD");
+    config.addAllowedMethod("GET");
+    config.addAllowedMethod("PUT");
+    config.addAllowedMethod("POST");
+    config.addAllowedMethod("DELETE");
+    config.addAllowedMethod("PATCH");
+    source.registerCorsConfiguration("/**", config);
+    return new CorsFilter(source);
+  }
 
 }
