+++ src/main/java/com/intellias/intellistart/interviewplanning/models/Booking.java
@@ -1,6 +1,5 @@
 package com.intellias.intellistart.interviewplanning.models;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.intellias.intellistart.interviewplanning.models.enums.Status;
 import java.time.LocalDateTime;
 import javax.persistence.Column;
@@ -14,7 +13,6 @@ import javax.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.hibernate.annotations.DynamicUpdate;
 
 /**
  * Booking entity class.
