+++ src/main/java/com/intellias/intellistart/interviewplanning/models/security/FacebookUser.java
@@ -4,24 +4,23 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.ToString;
 
+/**
+ * Facebook User class.
+ */
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Setter
+@ToString
 public class FacebookUser {
 
   private String id;
   private String email;
+  @SuppressWarnings("all")
   private String first_name;
+  @SuppressWarnings("all")
   private String last_name;
 
-  @Override
-  public String toString() {
-    return "FacebookUser{" +
-        "email='" + email + '\'' +
-        ", first_name='" + first_name + '\'' +
-        ", last_name='" + last_name + '\'' +
-        '}';
-  }
 }
