+++ src/main/java/com/intellias/intellistart/interviewplanning/payload/FacebookLoginRequest.java
@@ -1,10 +1,16 @@
 package com.intellias.intellistart.interviewplanning.payload;
 
-import lombok.*;
-
-import javax.validation.constraints.NotBlank;
 import java.io.Serializable;
+import javax.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+/**
+ * Facebook login request body.
+ */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
