+++ src/main/java/com/intellias/intellistart/interviewplanning/services/CandidateService.java
@@ -2,10 +2,8 @@ package com.intellias.intellistart.interviewplanning.services;
 
 import com.intellias.intellistart.interviewplanning.models.CandidateSlot;
 import com.intellias.intellistart.interviewplanning.repositories.CandidateSlotRepository;
-import com.intellias.intellistart.interviewplanning.util.exceptions.CandidateNotFoundException;
 import com.intellias.intellistart.interviewplanning.util.exceptions.CandidateSlotNotFoundException;
 import java.util.List;
-import java.util.Optional;
 import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
