+++ src/main/java/com/intellias/intellistart/interviewplanning/util/exceptions/InternalServerException.java
@@ -4,6 +4,9 @@ package com.intellias.intellistart.interviewplanning.util.exceptions;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
+/**
+ * Internal server error exception.
+ */
 @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
 public class InternalServerException extends RuntimeException {
